[quant][refactor tests] Move test_qat_module into test_quantize_eager_qat Summary:
Test Plan:
python test/test_quantization.py TestConvBNQATModule

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
